refactor: split out semantic and add tests
